\name{Weakest_link-package}
\alias{Weakest_link-package}
\alias{Weakest_link}
\docType{package}
\title{
\packageTitle{Weakest_link}
}
\description{
\packageDescription{Weakest_link}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Weakest_link}
\packageIndices{Weakest_link}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Weakest_link}

Maintainer: \packageMaintainer{Weakest_link}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
