\name{WeakestLink-package}
\alias{WeakestLink-package}
\alias{WeakestLink}
\docType{package}
\title{
\packageTitle{WeakestLink}
}
\description{
\packageDescription{WeakestLink}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{WeakestLink}
\packageIndices{WeakestLink}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{WeakestLink}

Maintainer: \packageMaintainer{WeakestLink}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
